// Frontend/cijene-frontend/src/types/api.ts
export interface Product {
  id: string;
  name: string;
  brand?: string;
  quantity?: string;
  unit?: string;
  ean?: string;
  chain_code?: string;
  chain?: string;
  category?: string;
  description?: string;
}
// ...
export interface Price {
  product_id: string;
  store_id: string;
  chain: string;
  price: number;
  special_price?: number;
  currency: string;
  date: string;
  unit?: string;
}